requirements.txt
flask[async]
mysql-connector-python==9.4.0
python-dotenv